+++ cmd/root.go
@@ -43,7 +43,7 @@ var rootCmd = &cobra.Command{
 	Use:           "safira",
 	Short:         "O Safira é uma ferramenta de auxílio ao Vertigo iPaaS",
 	Long:          "O Safira é uma ferramenta para auxiliar os desenvolvedores no Vertigo iPaaS",
-	Version:       "v0.0.1-beta.2",
+	Version:       "v0.0.2",
 	SilenceUsage:  true,
 	SilenceErrors: true,
 }